feat: track Astro Nvim config
I'd had this in a separate repo becuase I couldn't figure out how to
track a subdirectory of another repo in che
oi. Then I found out you could move the user config from `.config/nvim`
to `.config/astronvim`, so I did that.

Lost the history so far but not a big deal.